+++ back-end/youandi/src/main/java/com/ssafy/youandi/service/UserService.java
@@ -5,6 +5,6 @@ import com.ssafy.youandi.entity.user.SignupRequest;
 
 public interface UserService {
     //회원가입
-    public int singUp(SignupRequest signupRequest) throws Exception;
+    public Long singUp(SignupRequest signupRequest) throws Exception;
 
 }
